# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Vector Portal API
  version: 1.0.0
  description: |
    Document-workflow portal for multi-role approval chains.
    Manages events (document packages), their sections, and a
    role-based submit/approve/return pipeline with file attachments,
    inline comments, and a completed-document library.

    **Roles** (highest to lowest): ADMIN, PROTOCOL, DEPUTY, SUPERVISOR,
    SUPER_COLLABORATOR, COLLABORATOR, ANALYST.
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT returned by POST /api/auth/login (24 h expiry)
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          enum: [true]

    # ── Auth ──────────────────────────────────────────────────────────
...
    DocumentSubmitterRole:
      type: string
      enum: [DEPUTY, SUPERVISOR, SUPER_COLLABORATOR]
...
    SectionSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        sortOrder:
          type: integer
        departmentIds:
          type: array
          items:
            type: integer
    SectionInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          maxLength: 500
        departmentIds:
          type: array
          items:
            type: integer

    # ── Workflow ───────────────────────────────────────────────────────
...
    Template:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        documentSubmitterRole:
          $ref: '#/components/schemas/DocumentSubmitterRole'
        curatorRequired:
          type: boolean
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
        createdById:
          type: integer
          nullable: true
        createdByName:
          type: string
        sections:
          type: array
          items:
            $ref: '#/components/schemas/SectionSummary'

    # ── Admin ─────────────────────────────────────────────────────────
...
paths:
...
  /api/events/{id}/sections:
    post:
      tags: [events]
      summary: Add a section to an existing event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SectionInput'
      responses:
        '201':
          description: Section added
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  success:
                    type: boolean
...
  /api/sections:
    get:
      tags: [sections]
      summary: List sections for an event
      parameters:
        - name: event_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Array of sections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SectionSummary'
  /api/sections/{id}/label:
    patch:
      tags: [sections]
      summary: Rename a section
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                  maxLength: 300
      responses:
        '200':
          description: Section renamed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # ── Workflow ──────────────────────────────────────────────────────
...
  /api/templates:
    get:
      tags: [templates]
      summary: List event templates (defaults + own)
      responses:
        '200':
          description: Array of templates with sections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Template'
    post:
      tags: [templates]
      summary: Create an event template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 300
                documentSubmitterRole:
                  $ref: '#/components/schemas/DocumentSubmitterRole'
                curatorRequired:
                  type: boolean
                sections:
                  type: array
                  maxItems: 100
                  items:
                    $ref: '#/components/schemas/SectionInput'
      responses:
        '201':
          description: Template created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  success:
                    type: boolean
  /api/templates/{id}:
    delete:
      tags: [templates]
      summary: Delete own template (not default)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Template deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Template not found or cannot be deleted

  # ── Admin ─────────────────────────────────────────────────────────
...